% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genrets.R
\name{genrets}
\alias{genrets}
\title{Create a tibble or zoo object of returns using tq_get}
\usage{
genrets(
  tickers,
  start_date,
  freq = "daily",
  ret_type = "log",
  output = "tibble",
  long = TRUE
)
}
\arguments{
\item{tickers}{A list of yahoo finance tickers. \code{list}}

\item{start_date}{the first observation date. \code{character}}

\item{freq}{daily (default), 'monthly'. \code{character}}

\item{ret_type}{log (default), 'rel', or 'abs'. \code{character}}

\item{output}{'tibble' (default), 'zoo'. \code{character}}

\item{long}{TRUE (default) to return a long tibble FALSE for wide. Zoo objects are wide. \code{character}}
}
\value{
Long \code{tibble} (default), Wide \code{tibble}, \code{Zoo} object.
}
\description{
This function takes ticker(s), start date, return type, and output type as inputs and generates either a tibble or zoo object
using the tidyquant, tidy, and zoo libraries.
}
\examples{
tickers = c('SPY', 'QQQ')
rets_long <- genrets(tickers, '2020-01-01')

}
\author{
Mitch Greer
}
