% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retprofile.R
\name{retprofile}
\alias{retprofile}
\title{Generate a visual profile of returns across metrics like rolling SD, rolling corr, and cumulative returns}
\usage{
retprofile(long_df, window, ret_type = "log")
}
\arguments{
\item{long_df}{A long dataframe \code{data.frame}}

\item{window}{the window for rolling measurements \code{numeric}}

\item{ret_type}{the return type to calculate cumulative returns from \code{character}}
}
\description{
This function takes a long dataframe and outputs several plotly graphs for standard deviation, correlation, and cumulative return
using the slider and plotly libraries. DF columns names be date, symbol, value
}
\examples{
\dontrun{set.seed(123)
dates <- seq.Date(as.Date("2020-01-01"), as.Date("2020-12-31"), by = "day")
tickers <- c("TICKER_A", "TICKER_B")
data <- data.frame(
  date = rep(dates, each = length(tickrs)),
  ticker = rep(tickers, times = length(dates)),
  return = rnorm(length(dates) * length(tickers), mean = 0, sd = 0.04)
  )
retprofile(data, window = 30, ret_type = 'log')}
}
\author{
Travis Nowak
}
